+++ src/entities/abstractions/element.py
@@ -2,8 +2,7 @@ from typing import List
 
 from pygame import Rect, Surface, time
 
-from src.utils.camera import Camera
-from src.utils.classes import Position
+from src.utils import Camera, Position
 from src.utils.constants import ANIMATION_INTERVAL, INIT_IMAGE_INDEX
 
 from ..interfaces import IEntity
