+++ src/utils/__init__.py
@@ -1,3 +1,4 @@
+from .camera import Camera
 from .classes import Position, Singleton, Size
 
-__all__ = ["Position", "Size", "Singleton"]
+__all__ = ["Position", "Size", "Singleton", "Camera"]
